Kpm system Issue #70 enhancements
added couple of options including one to show on the terminal the data as JSON. now for the java properties it will check if there it is a running apache tomcat proccess; if there it is one, it will be used to get the properties. If no proccess is found then it will get the properies from the installed java. Tested on windows and Ubuntu Linux